requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.28.0
datasets>=2.18.0
bitsandbytes>=0.43.0

# Geometry & Analysis
scikit-learn>=1.4.0
umap-learn>=0.5.6
numpy>=1.26.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
